feat: write with zustand about change actionType logic
