+++ tsconfig.json
@@ -30,7 +30,8 @@
       "@themes": ["themes"],
       "@hooks": ["hooks"],
       "@routes": ["routes"],
-      "@services": ["services"]
+      "@services": ["services"],
+      "@stores": ["stores"]
     }
   },
   "include": ["src"],
